API Reference

Functions exported:

Discontinuity.PVI — Function

PVI(B, n, ::Val{dim})

Compute the Partial Variance of Increments (PVI) for a time series B along dimension dim with lag n. Standard deviations σi of magnetic field increments are computed every time units.

\[\mathrm{PVI}_{t, τ} = \frac{|Δ \mathbf{B}(t, τ)|}{\sqrt{\langle | Δ \mathbf{B}(t, τ) |^2\rangle}}\]

Returns a Vector of length N - τ containing PVI values.

Discontinuity.compute_anisotropy_params! — Method

computeanisotropyparams!(df, :ion => (:Tppara, :Tpperp), :electron => (:Tepara, :Teperp))

Discontinuity.detect_variance — Method

detect_variance(f, tmin, tmax, tau; split=1, kwargs...)

Apply detect_variance on a time series f(tmin, tmax)

Optional splitting into multiple chunks to improve performance and memory efficiency.

Discontinuity.detect_variance — Method

detect_variance(data, times, period, sparse_num, d; kwargs...)
detect_variance(data::AbstractDimArray, period, sparse_num; dim=Ti, kwargs...)

Detect discontinuities based on variance analysis.

Keyword arguments:

n=2: Number of periods to split the time series into

std_threshold=2: Threshold for standard deviation

fluc_threshold=1: Threshold for index fluctuation

diff_threshold=0.1: Threshold for difference

Discontinuity.ids_finder — Method

ids_finder(data, tau, [V, n]; kwargs...)

Find the discontinuities in the time series data with the given tau.

Provide additional arguments V and n to calculate the features of the events.

Discontinuity.ids_finder — Method

ids_finder(f, tmin, tmax, tau, args...; split=1, kwargs...)

Apply ids_finder on a time series f(tmin, tmax)

Optional splitting into multiple chunks to improve performance and memory efficiency.

Discontinuity.load — Method

load(path)

Load the data from the given path

Types exported:

Discontinuity.HyperbolicTangentFit — Type

HyperbolicTangentFit{T}

A hyperbolic tangent fit.

\[y = A tanh((x - μ) / σ) + B\]